OGM-366 Move database identifier class
