+++ core/src/main/java/org/hibernate/ogm/remote/impl/DatabaseIdentifier.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.couchdb.util.impl;
+package org.hibernate.ogm.remote.impl;
 
 import java.net.MalformedURLException;
 import java.net.URI;
@@ -12,7 +12,7 @@ import java.net.URISyntaxException;
 import java.net.URL;
 
 /**
- * Provides all information required to connect to a CouchDB database.
+ * Provides all information required to connect to a database.
  *
  * @author Andrea Boriero &lt;dreborier@gmail.com&gt;
  * @author Gunnar Morling
